gcp/requirements.txt
tensorflow
numpy
pillow
google-cloud-storage
functions-framework
flask
